fixed some compiler warnings to do with time formatting for the debug build

+++ accidental_irc.c
@@ -2322,7 +2322,7 @@ void server_privmsg_command(int server_index, char *tmp_buffer, int first_space,
 	if(!strcmp(tmp_nick,channel)){
 #ifdef DEBUG
 		char sys_call_buffer[BUFFER_SIZE];
-		sprintf(sys_call_buffer,"echo \"%ju <%s> %s\" | mail -s \"PM\" \"%s\"",time(NULL),nick,text,servers[server_index]->nick);
+		sprintf(sys_call_buffer,"echo \"%ju <%s> %s\" | mail -s \"PM\" \"%s\"",(uintmax_t)(time(NULL)),nick,text,servers[server_index]->nick);
 		system(sys_call_buffer);
 #endif
 		//set this to the last PM-ing user, so we can later reply if we so choose
@@ -2434,7 +2434,7 @@ void server_privmsg_command(int server_index, char *tmp_buffer, int first_space,
 #ifdef DEBUG
 		//take any desired additional steps upon ping here (could add notify-send or something, if desired)
 		char sys_call_buffer[BUFFER_SIZE];
-		sprintf(sys_call_buffer,"echo \"%ju ***<%s> %s\" | mail -s \"PING\" \"%s\"",time(NULL),nick,text,servers[server_index]->nick);
+		sprintf(sys_call_buffer,"echo \"%ju ***<%s> %s\" | mail -s \"PING\" \"%s\"",(uintmax_t)(time(NULL)),nick,text,servers[server_index]->nick);
 		system(sys_call_buffer);
 #endif
 		//audio output
